signed unsigned warning removal

+++ JAERO/oqpskdemodulator.cpp
@@ -250,7 +250,7 @@ qint64 OqpskDemodulator::writeData(const char *data, qint64 len)
     bool sendscatterpoints=false;
 
     const short *ptr = reinterpret_cast<const short *>(data);
-    for(int i=0;i<len/sizeof(short);i++)
+    for(int i=0;i<((int)(len/sizeof(short)));i++)
     {
         double dval=((double)(*ptr))/32768.0;
